# docs/swagger.yaml
basePath: /api
definitions:
  ds.Lamp:
    properties:
      id:
        type: integer
      image_url:
        type: string
      luminous_flux_lm:
        type: number
      power_w:
        type: number
      scattering_angle_deg:
        type: number
      title:
        type: string
    type: object
...
  handler.CreateLampRequest:
    properties:
      luminous_flux_lm:
        type: number
      power_w:
        type: number
      scattering_angle_deg:
        type: number
      title:
        type: string
    required:
    - luminous_flux_lm
    - power_w
    - scattering_angle_deg
    - title
    type: object
...
  handler.UpdateLampRequest:
    properties:
      luminous_flux_lm:
        type: number
      power_w:
        type: number
      scattering_angle_deg:
        type: number
      title:
        type: string
    type: object
...
host: localhost:8001
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://github.com/yourusername/dia-backend
  description: REST API для веб-приложения расчета освещения комнат
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Let There Be Light API
  version: "1.0"
paths:
  /lamps:
    get:
      consumes:
      - application/json
      description: Get list of all lamps with optional title filter
      parameters:
      - description: Filter by lamp title
        in: query
        name: title
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ds.Lamp'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List lamps
      tags:
      - lamps
    post:
      consumes:
      - application/json
      description: Create a new lamp (moderator only)
      parameters:
      - description: Lamp data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateLampRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ds.Lamp'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new lamp
      tags:
      - lamps
  /lamps/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a lamp (moderator only)
      parameters:
      - description: Lamp ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete lamp
      tags:
      - lamps
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific lamp
      parameters:
      - description: Lamp ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ds.Lamp'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get lamp by ID
      tags:
      - lamps
    put:
      consumes:
      - application/json
      description: Update lamp information (moderator only)
      parameters:
      - description: Lamp ID
        in: path
        name: id
        required: true
        type: integer
      - description: Lamp update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateLampRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update lamp
      tags:
      - lamps
  /lamps/{id}/draft:
    post:
      consumes:
      - application/json
      description: Add lamp to user's draft light request
      parameters:
      - description: Lamp ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add lamp to draft request
      tags:
      - lamps
  /lamps/{id}/image:
    post:
      consumes:
      - multipart/form-data
      description: Upload or replace lamp image (moderator only)
      parameters:
      - description: Lamp ID
        in: path
        name: id
        required: true
        type: integer
      - description: Lamp image file
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add lamp image
      tags:
      - lamps
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"